tcp connection management in hub
